Add restrictions when the user may leave the trip or change role

+++ src/main/kotlin/cz/cvut/fit/steuejan/travel/api/trip/controller/TripController.kt
@@ -11,6 +11,9 @@ import cz.cvut.fit.steuejan.travel.api.app.response.Status
 import cz.cvut.fit.steuejan.travel.api.app.response.Success
 import cz.cvut.fit.steuejan.travel.api.trip.document.model.DocumentOverview
 import cz.cvut.fit.steuejan.travel.api.trip.document.response.DocumentOverviewListResponse
+import cz.cvut.fit.steuejan.travel.api.trip.exception.CannotChangeRoleException
+import cz.cvut.fit.steuejan.travel.api.trip.exception.CannotChangeRoleToMyselfException
+import cz.cvut.fit.steuejan.travel.api.trip.exception.CannotLeaveException
 import cz.cvut.fit.steuejan.travel.api.trip.model.ChangeRole
 import cz.cvut.fit.steuejan.travel.api.trip.model.TripInvitation
 import cz.cvut.fit.steuejan.travel.api.trip.model.TripUser
@@ -32,7 +35,7 @@ class TripController(daoFactory: DaoFactory) : AbstractTripController(daoFactory
     }
 
     suspend fun deleteTrip(userId: Int, tripId: Int): Response {
-        if (daoFactory.tripUserDao.findConnection(userId, tripId)?.role != UserRole.ADMIN) {
+        if (getUserRole(userId, tripId) != UserRole.ADMIN) {
             throw ForbiddenException(FailureMessages.DELETE_TRIP_PROHIBITED)
         }
 
@@ -57,11 +60,20 @@ class TripController(daoFactory: DaoFactory) : AbstractTripController(daoFactory
     }
 
     suspend fun invite(userId: Int, invitation: TripInvitation): Response {
-        editOrThrow(userId, invitation.tripId)
+        val userRole = getUserRole(userId, invitation.tripId)
+
+        if (!userRole.canEdit()) {
+            throw ForbiddenException(FailureMessages.INVITE_PROHIBITED)
+        }
 
         with(invitation) {
             val user = daoFactory.userDao.findByUsername(username)
                 ?: throw NotFoundException(FailureMessages.USER_NOT_FOUND)
+
+            //editor cannot invite user who will have admin rights
+            if (userRole == UserRole.EDITOR && role == UserRole.ADMIN) {
+                throw ForbiddenException(FailureMessages.INVITE_EDITOR_PROHIBITED)
+            }
             daoFactory.tripUserDao.addConnection(user.id, tripId, role)
         }
 
@@ -90,15 +102,28 @@ class TripController(daoFactory: DaoFactory) : AbstractTripController(daoFactory
     }
 
     suspend fun leave(userId: Int, tripId: Int): Response {
-        if (!daoFactory.tripUserDao.removeConnection(userId, tripId)) {
-            throw NotFoundException(FailureMessages.USER_TRIP_NOT_FOUND)
+        val dao = daoFactory.tripUserDao
+
+        val isAdmin = getUserRole(userId, tripId) == UserRole.ADMIN
+
+        //user can leave trip if he is alone or there is another admin
+        if (!isAdmin || dao.countUsersInTrip(tripId) == 1L || dao.countAdminsInTrip(tripId) >= 2L) {
+            if (!daoFactory.tripUserDao.removeConnection(userId, tripId)) {
+                throw ForbiddenException(FailureMessages.USER_TRIP_NOT_FOUND)
+            }
+        } else {
+            throw CannotLeaveException()
         }
+
         return Success(Status.NO_CONTENT)
     }
 
     suspend fun changeRole(userId: Int, tripId: Int, changeRole: ChangeRole): Response {
-        if (daoFactory.tripUserDao.findConnection(userId, tripId)?.role != UserRole.ADMIN) {
-            throw ForbiddenException(FailureMessages.EDIT_CHANGE_ROLE_PROHIBITED)
+        val dao = daoFactory.tripUserDao
+        val userRole = getUserRole(userId, tripId)
+
+        if (userRole != UserRole.ADMIN) {
+            throw CannotChangeRoleException()
         }
 
         //remove user if no role set
@@ -106,7 +131,12 @@ class TripController(daoFactory: DaoFactory) : AbstractTripController(daoFactory
             return leave(changeRole.whomId, tripId)
         }
 
-        if (!daoFactory.tripUserDao.changeRole(changeRole.whomId, tripId, changeRole.newRole)) {
+        //at least one admin must stay in the trip
+        if (dao.countAdminsInTrip(tripId) == 1L && userId == changeRole.whomId) {
+            throw CannotChangeRoleToMyselfException()
+        }
+
+        if (!dao.changeRole(changeRole.whomId, tripId, changeRole.newRole)) {
             throw NotFoundException(FailureMessages.USER_TRIP_NOT_FOUND)
         }
